backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=5.2.4",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
]
